requirements.txt
contentful_management
python-frontmatter